Added new forms for bill and vote lookup: can now filter bill form by bill types. New vote form which can filter by date and chamber. New Bill form to select a specific bill by congress, type and number.
Added loose thread for future prediction of a senator's vote on a bill. Currently function is implemented but user interface needed.

Unified value names for urls and views. For example, congress_num now is the uniform name for the congrss number, before it was a mix of congress_num, congress, congress_id, etc.